Get_ReverseShell/requirements.txt
python-nmap==0.6.1
requests==2.18.4
